Sam's code didnt compile fixed it.
A good amount of misspelled words as well as minor chagnes to the code
like missing "new" to make a class or small items of that nature. Could
have been fixed with a compile and check.

+++ java/FightItem.java
@@ -2,10 +2,10 @@ import java.util.Random;
 
 public abstract class FightItem extends Usable
 {
-   protected int damage;
+   protected double damage;
    protected double acuracy;
    
-   public int use()
+   public double use()
    {
       uses--;
       Random myRand = new Random();
